+++ src/server/services/index.ts
@@ -3,6 +3,7 @@ export * from './api/index';
 export * from './calendar/index';
 export * from './dbo/index';
 export * from './log-service/index';
+export * from './network/index';
 export * from './room-service/index';
 export * from './Sonos/index';
 export * from './Telegram/index';