+++ CHANGELOG.md
@@ -4,8 +4,7 @@
   Placeholder for the next version (at the beginning of the line):
   ## **WORK IN PROGRESS**
 -->
-
-## **WORK IN PROGRESS**
+## 0.1.42-alpha.8 (2022-07-14)
 
 * (theimo1221) Implement Router
 * (theimo1221) Add Network reconnect to daikin device